backend/web-back/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
django-cors-headers==3.11.0
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djoser==2.0.1
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
